separate StoryPromo's custom hook into colocated file
